I'm sorry, @Nikolaj64, but web works like that.
You shouldn't disable select cursor just because it looks better
that way (hint - it doesn't). This is something that browser should
decide, not website.

In desktop browsers, select cursor is hint that you can select the
text. While it could work without it, why break that just to break
that?