Remove layout maker
Since its introduction, it was used to create awful layouts, that
caused many ABXD boards administrators to remove such layouts, and
not enable layout maker. As I don't see point of having a feature
that everyone hates, I decided to outright remove it, in order to
remove feature that frustrated everybody, and the traces of it are
still visible, even when most ABXD boards don't have it enabled.